Initial commit for changing icon depending on time
This is a new feature that changes the icon depending on the time that
was spent in the pomodoro. The icons were taken from gpomo project which
I used in the past and I liked the idea. Note that this removes the
possibility for the user to specify a custom icon through beautiful. The
old behaviour had a bug so this shouldn't bother anyone and they will
probably just open a feature request for it as needed